+++ src/Polly.Contrib.Simmy.Specs/Latency/InjectLatencyAsyncWithOptionsSpecs.cs
@@ -6,7 +6,6 @@ using FluentAssertions;
 using Polly.Contrib.Simmy.Utilities;
 using Polly.Utilities;
 using Xunit;
-using Polly.Contrib.Simmy.Latency.Options;
 using Polly.Timeout;
 
 namespace Polly.Contrib.Simmy.Specs.Latency
