+++ src/Polly.Contrib.Simmy.Specs/Latency/InjectLatencyTResultWithOptionsSpecs.cs
@@ -5,7 +5,6 @@ using Polly.Contrib.Simmy.Specs.Helpers;
 using Polly.Contrib.Simmy.Utilities;
 using Polly.Utilities;
 using Xunit;
-using Polly.Contrib.Simmy.Latency.Options;
 using System.Threading;
 using Polly.Timeout;
