+++ src/Polly.Contrib.Simmy.Specs/Outcomes/InjectFaultAsyncSpecs.cs
@@ -6,7 +6,7 @@ using Polly.Contrib.Simmy.Utilities;
 using Polly.Utilities;
 using Xunit;
 
-namespace Polly.Contrib.Simmy.Specs.Fault
+namespace Polly.Contrib.Simmy.Specs.Outcomes
 {
     [Collection(Helpers.Constants.AmbientContextDependentTestCollection)]
     [Obsolete]
