+++ src/Polly.Contrib.Simmy.Specs/Outcomes/InjectFaultSpecs.cs
@@ -3,14 +3,14 @@ using System.Threading;
 using FluentAssertions;
 using Polly.Contrib.Simmy.Utilities;
 using Xunit;
-using Polly.Contrib.Simmy.Fault.Options;
 
-namespace Polly.Contrib.Simmy.Specs.Fault
+namespace Polly.Contrib.Simmy.Specs.Outcomes
 {
     [Collection(Helpers.Constants.AmbientContextDependentTestCollection)]
-    public class InjectFaultWithOptionsSpecs : IDisposable
+    [Obsolete]
+    public class InjectFaultSpecs : IDisposable
     {
-        public InjectFaultWithOptionsSpecs()
+        public InjectFaultSpecs()
         {
             ThreadSafeRandom_LockOncePerThread.NextDouble = () => 0.5;
         }
@@ -26,12 +26,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string exceptionMessage = "exceptionMessage";
             Exception fault = new Exception(exceptionMessage);
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(0.6)
-                    .Enabled()
-            );
+            var policy = MonkeyPolicy.InjectFault(fault, 0.6, () => true);
 
             Boolean executed = false;
             policy.Invoking(x => x.Execute(() => { executed = true; }))
@@ -44,12 +39,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_Context_Free_Enabled_Should_execute_user_delegate()
         {
             Exception fault = new Exception("test");
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(0.3)
-                    .Enabled()
-            );
+            var policy = MonkeyPolicy.InjectFault(fault, 0.3, () => true);
 
             Boolean executed = false;
             policy.Invoking(x => x.Execute(() => { executed = true; }))
@@ -62,12 +52,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_Context_Free_Enabled_Should_execute_user_delegate_not_throw_if_injected_fault_is_permitted_null()
         {
             Exception fault = null;
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(0.6)
-                    .Enabled()
-            );
+            var policy = MonkeyPolicy.InjectFault(fault, 0.6, () => true);
 
             Boolean executed = false;
             policy.Invoking(x => x.Execute(() => { executed = true; }))
@@ -83,13 +68,15 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_not_execute_user_delegate()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = true };
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(new Exception("test"))
-                    .InjectionRate(0.6)
-                    .EnabledWhen((ctx, ct) => ((bool)ctx["ShouldFail"]))
-            );
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            var policy = MonkeyPolicy.InjectFault(
+                new Exception("test"),
+                0.6,
+                (ctx, ct) =>
+                {
+                    return ((bool)ctx["ShouldFail"]);
+                });
 
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldThrowExactly<Exception>();
@@ -102,13 +89,15 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             Exception fault = new Exception("test");
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = true };
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(0.4)
-                    .EnabledWhen((ctx, ct) => ((bool)ctx["ShouldFail"]))
-            );
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            var policy = MonkeyPolicy.InjectFault(
+                new Exception("test"),
+                0.4,
+                (ctx, ct) =>
+                {
+                    return ((bool)ctx["ShouldFail"]);
+                });
 
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldNotThrow<Exception>();
@@ -121,13 +110,15 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             Exception fault = new Exception("test");
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = false };
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(0.6)
-                    .EnabledWhen((ctx, ct) => ((bool)ctx["ShouldFail"]))
-            );
+            Context context = new Context();
+            context["ShouldFail"] = false;
+            var policy = MonkeyPolicy.InjectFault(
+                new Exception("test"),
+                0.6,
+                (ctx, ct) =>
+                {
+                    return ((bool)ctx["ShouldFail"]);
+                });
 
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldNotThrow<Exception>();
@@ -147,13 +138,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<Context, CancellationToken, Exception> fault = (ctx, ct) => new Exception();
             Func<Context, CancellationToken, double> injectionRate = (ctx, ct) => -0.1;
             Func<Context, CancellationToken, bool> enabled = (ctx, ct) => true;
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldThrowExactly<ArgumentOutOfRangeException>();
 
@@ -169,13 +154,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<Context, CancellationToken, Exception> fault = (ctx, ct) => new Exception();
             Func<Context, CancellationToken, double> injectionRate = (ctx, ct) => 1.01;
             Func<Context, CancellationToken, bool> enabled = (ctx, ct) => true;
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldThrowExactly<ArgumentOutOfRangeException>();
 
@@ -191,13 +170,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<Context, CancellationToken, Exception> fault = (ctx, ct) => new Exception();
             Func<Context, CancellationToken, double> injectionRate = (ctx, ct) => 0.6;
             Func<Context, CancellationToken, bool> enabled = (ctx, ct) => true;
-
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldThrowExactly<Exception>();
 
@@ -208,13 +181,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_not_execute_user_delegate_with_all_context_set()
         {
             Boolean executed = false;
+            Context context = new Context();
             string failureMessage = "Failure Message";
-            Context context = new Context
-            {
-                ["ShouldFail"] = true,
-                ["Message"] = failureMessage,
-                ["InjectionRate"] = 0.6
-            };
+            context["ShouldFail"] = true;
+            context["Message"] = failureMessage;
+            context["InjectionRate"] = 0.6;
 
             Func<Context, CancellationToken, Exception> fault = (ctx, ct) =>
             {
@@ -241,12 +212,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return ((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
             policy.Invoking(x => x.Execute((ctx) => { executed = true; }, context))
                 .ShouldThrowExactly<InvalidOperationException>().WithMessage(failureMessage);
 
@@ -261,12 +227,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context
-            {
-                ["ShouldFail"] = true,
-                ["Message"] = failureMessage,
-                ["InjectionRate"] = 0.6
-            };
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            context["Message"] = failureMessage;
+            context["InjectionRate"] = 0.6;
 
             Func<Context, CancellationToken, Exception> fault = (ctx, cts) =>
             {
@@ -295,12 +259,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return (bool)ctx["ShouldFail"];
             };
 
-            var policy = MonkeyPolicy.InjectFault(with =>
-                with.Fault(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
             using (CancellationTokenSource cts = new CancellationTokenSource())
             {
                 cts.Cancel();
@@ -317,12 +276,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context
-            {
-                ["ShouldFail"] = true,
-                ["Message"] = failureMessage,
-                ["InjectionRate"] = 0.6
-            };
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            context["Message"] = failureMessage;
+            context["InjectionRate"] = 0.6;
 
             Func<Context, CancellationToken, Exception> fault = (ctx, cts) =>
             {
@@ -354,11 +311,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return (bool)ctx["ShouldFail"];
                 };
 
-                var policy = MonkeyPolicy.InjectFault(with =>
-                    with.Fault(fault)
-                        .InjectionRate(injectionRate)
-                        .EnabledWhen(enabled)
-                );
+                var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
 
                 policy.Invoking(x => x.Execute((ctx, ct) => { executed = true; }, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -372,12 +325,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context
-            {
-                ["ShouldFail"] = true,
-                ["Message"] = failureMessage,
-                ["InjectionRate"] = 0.6
-            };
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            context["Message"] = failureMessage;
+            context["InjectionRate"] = 0.6;
 
             Func<Context, CancellationToken, Exception> fault = (ctx, cts) =>
             {
@@ -409,11 +360,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return rate;
                 };
 
-                var policy = MonkeyPolicy.InjectFault(with =>
-                    with.Fault(fault)
-                        .InjectionRate(injectionRate)
-                        .EnabledWhen(enabled)
-                );
+                var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
 
                 policy.Invoking(x => x.Execute((ctx, ct) => { executed = true; }, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -427,12 +374,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context
-            {
-                ["ShouldFail"] = true,
-                ["Message"] = failureMessage,
-                ["InjectionRate"] = 0.6
-            };
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            context["Message"] = failureMessage;
+            context["InjectionRate"] = 0.6;
 
             using (CancellationTokenSource cts = new CancellationTokenSource())
             {
@@ -464,11 +409,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return new Exception();
                 };
 
-                var policy = MonkeyPolicy.InjectFault(with =>
-                    with.Fault(fault)
-                        .InjectionRate(injectionRate)
-                        .EnabledWhen(enabled)
-                );
+                var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
 
                 policy.Invoking(x => x.Execute((ctx, ct) => { executed = true; }, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -478,6 +419,5 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         }
 
         #endregion
-
     }
 }
