+++ src/Polly.Contrib.Simmy.Specs/Outcomes/InjectFaultTResultAsyncSpecs.cs
@@ -5,14 +5,14 @@ using FluentAssertions;
 using Polly.Contrib.Simmy.Specs.Helpers;
 using Polly.Contrib.Simmy.Utilities;
 using Xunit;
-using Polly.Contrib.Simmy.Fault.Options;
 
-namespace Polly.Contrib.Simmy.Specs.Fault
+namespace Polly.Contrib.Simmy.Specs.Outcomes
 {
-    [Collection(Helpers.Constants.AmbientContextDependentTestCollection)]
-    public class InjectFaultTResultAsyncWithOptionsSpecs : IDisposable
+    [Collection(Constants.AmbientContextDependentTestCollection)]
+    [Obsolete]
+    public class InjectFaultTResultAsyncSpecs : IDisposable
     {
-        public InjectFaultTResultAsyncWithOptionsSpecs()
+        public InjectFaultTResultAsyncSpecs()
         {
             ThreadSafeRandom_LockOncePerThread.NextDouble = () => 0.5;
         }
@@ -31,12 +31,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Boolean executed = false;
             Func<Task<ResultPrimitive>> actionAsync = () => { executed = true; return Task.FromResult(ResultPrimitive.Good); };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault<ResultPrimitive>(fault)
-                    .InjectionRate(0.6)
-                    .Enabled()
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, () => true);
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync))
                 .ShouldThrowExactly<Exception>()
                 .WithMessage(exceptionMessage);
@@ -51,12 +46,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Boolean executed = false;
             Func<Task<ResultPrimitive>> actionAsync = () => { executed = true; return Task.FromResult(ResultPrimitive.Good); };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault<ResultPrimitive>(fault)
-                    .InjectionRate(0.3)
-                    .Enabled()
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.3, () => true);
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync))
                 .ShouldNotThrow<Exception>();
             executed.Should().BeTrue();
@@ -68,17 +58,20 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_not_execute_user_delegate_async()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = true };
+            Context context = new Context();
+            context["ShouldFail"] = true;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true; return Task.FromResult(ResultPrimitive.Good);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault(new Exception())
-                    .InjectionRate(0.6)
-                    .EnabledWhen(async (ctx, ct) => await Task.FromResult((bool)ctx["ShouldFail"]))
-            );
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(
+                new Exception(),
+                0.6,
+                async (ctx, ct) =>
+                {
+                    return await Task.FromResult((bool)ctx["ShouldFail"]);
+                });
 
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context)).ShouldThrowExactly<Exception>();
             executed.Should().BeFalse();
@@ -88,17 +81,20 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_execute_user_delegate_async()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = true };
+            Context context = new Context();
+            context["ShouldFail"] = true;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true; return Task.FromResult(ResultPrimitive.Good);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault(new Exception())
-                    .InjectionRate(0.4)
-                    .EnabledWhen(async (ctx, ct) => await Task.FromResult((bool)ctx["ShouldFail"]))
-            );
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(
+                new Exception(),
+                0.4,
+                async (ctx, ct) =>
+                {
+                    return await Task.FromResult((bool)ctx["ShouldFail"]);
+                });
 
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldNotThrow<Exception>();
@@ -110,17 +106,20 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_execute_user_delegate_async_with_enabled_lambda_return_false()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = false };
+            Context context = new Context();
+            context["ShouldFail"] = false;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true; return Task.FromResult(ResultPrimitive.Good);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault(new Exception())
-                    .InjectionRate(0.4)
-                    .EnabledWhen(async (ctx, ct) => await Task.FromResult((bool)ctx["ShouldFail"]))
-            );
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(
+                new Exception(),
+                0.4,
+                async (ctx, ct) =>
+                {
+                    return await Task.FromResult((bool)ctx["ShouldFail"]);
+                });
 
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldNotThrow<Exception>();
@@ -134,7 +133,8 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_not_execute_user_delegate_async_with_default_values()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = true };
+            Context context = new Context();
+            context["ShouldFail"] = true;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -144,13 +144,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<Context, CancellationToken, Task<Exception>> fault = (ctx, cts) => Task.FromResult(new Exception());
             Func<Context, CancellationToken, Task<double>> injectionRate = (ctx, ct) => Task.FromResult(0.6);
             Func<Context, CancellationToken, Task<bool>> enabled = (ctx, ct) => Task.FromResult(true);
-
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault<ResultPrimitive>(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldThrowExactly<Exception>();
             executed.Should().BeFalse();
@@ -161,12 +155,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context
-            {
-                ["ShouldFail"] = true,
-                ["Message"] = failureMessage,
-                ["InjectionRate"] = 0.6
-            };
+            Context context = new Context();
+            context["ShouldFail"] = true;
+            context["Message"] = failureMessage;
+            context["InjectionRate"] = 0.6;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -200,12 +192,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Fault(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldThrowExactly<InvalidOperationException>();
             executed.Should().BeFalse();
@@ -226,12 +213,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
 
             ResultPrimitive fault = ResultPrimitive.Fault;
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(0.6)
-                    .Enabled()
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, () => true);
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -249,12 +231,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
 
             ResultPrimitive fault = ResultPrimitive.Fault;
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(0.4)
-                    .Enabled()
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.4, () => true);
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -264,7 +241,8 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public async Task InjectFault_With_Context_Enabled_Should_Return_Fault_async()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = true };
+            Context context = new Context();
+            context["ShouldFail"] = true;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -277,12 +255,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(0.6)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, enabled);
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -292,7 +265,8 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public async Task InjectFault_With_Context_Enabled_Should_Not_Return_Fault_async()
         {
             Boolean executed = false;
-            Context context = new Context { ["ShouldFail"] = false };
+            Context context = new Context();
+            context["ShouldFail"] = false;
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -305,12 +279,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(0.6)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, enabled);
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -320,7 +289,8 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public async Task InjectFault_With_Context_InjectionRate_Should_Return_Fault_async()
         {
             Boolean executed = false;
-            Context context = new Context { ["InjectionRate"] = 0.6 };
+            Context context = new Context();
+            context["InjectionRate"] = 0.6;
 
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
@@ -349,12 +319,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult(true);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -394,12 +359,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult(true);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -450,12 +410,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                with.Result(fault)
-                    .InjectionRate(injectionRate)
-                    .EnabledWhen(enabled)
-            );
-
+            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
             using (CancellationTokenSource cts = new CancellationTokenSource())
             {
                 cts.Cancel();
@@ -512,11 +467,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return Task.FromResult((bool)ctx["ShouldFail"]);
                 };
 
-                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                    with.Result(fault)
-                        .InjectionRate(injectionRate)
-                        .EnabledWhen(enabled)
-                );
+                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
 
                 policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -570,11 +521,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return Task.FromResult(rate);
                 };
 
-                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                    with.Result(fault)
-                        .InjectionRate(injectionRate)
-                        .EnabledWhen(enabled)
-                );
+                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
 
                 policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -628,11 +575,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return Task.FromResult(new Exception());
                 };
 
-                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(with =>
-                    with.Result(fault)
-                        .InjectionRate(injectionRate)
-                        .EnabledWhen(enabled)
-                );
+                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
 
                 policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
