+++ src/Polly.Contrib.Simmy.Specs/Outcomes/InjectFaultTResultAsyncWithOptionsSpecs.cs
@@ -2,18 +2,16 @@
 using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
-using Polly;
 using Polly.Contrib.Simmy.Specs.Helpers;
 using Polly.Contrib.Simmy.Utilities;
 using Xunit;
 
-namespace Polly.Contrib.Simmy.Specs.Fault
+namespace Polly.Contrib.Simmy.Specs.Outcomes
 {
-    [Collection(Constants.AmbientContextDependentTestCollection)]
-    [Obsolete]
-    public class InjectFaultTResultAsyncSpecs : IDisposable
+    [Collection(Helpers.Constants.AmbientContextDependentTestCollection)]
+    public class InjectFaultTResultAsyncWithOptionsSpecs : IDisposable
     {
-        public InjectFaultTResultAsyncSpecs()
+        public InjectFaultTResultAsyncWithOptionsSpecs()
         {
             ThreadSafeRandom_LockOncePerThread.NextDouble = () => 0.5;
         }
@@ -32,7 +30,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Boolean executed = false;
             Func<Task<ResultPrimitive>> actionAsync = () => { executed = true; return Task.FromResult(ResultPrimitive.Good); };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, () => true);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault<ResultPrimitive>(fault)
+                    .InjectionRate(0.6)
+                    .Enabled()
+            );
+
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync))
                 .ShouldThrowExactly<Exception>()
                 .WithMessage(exceptionMessage);
@@ -47,7 +50,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Boolean executed = false;
             Func<Task<ResultPrimitive>> actionAsync = () => { executed = true; return Task.FromResult(ResultPrimitive.Good); };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.3, () => true);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault<ResultPrimitive>(fault)
+                    .InjectionRate(0.3)
+                    .Enabled()
+            );
+
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync))
                 .ShouldNotThrow<Exception>();
             executed.Should().BeTrue();
@@ -59,20 +67,17 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_not_execute_user_delegate_async()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true; return Task.FromResult(ResultPrimitive.Good);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(
-                new Exception(),
-                0.6,
-                async (ctx, ct) =>
-                {
-                    return await Task.FromResult((bool)ctx["ShouldFail"]);
-                });
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault(new Exception())
+                    .InjectionRate(0.6)
+                    .EnabledWhen(async (ctx, ct) => await Task.FromResult((bool)ctx["ShouldFail"]))
+            );
 
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context)).ShouldThrowExactly<Exception>();
             executed.Should().BeFalse();
@@ -82,20 +87,17 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_execute_user_delegate_async()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true; return Task.FromResult(ResultPrimitive.Good);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(
-                new Exception(),
-                0.4,
-                async (ctx, ct) =>
-                {
-                    return await Task.FromResult((bool)ctx["ShouldFail"]);
-                });
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault(new Exception())
+                    .InjectionRate(0.4)
+                    .EnabledWhen(async (ctx, ct) => await Task.FromResult((bool)ctx["ShouldFail"]))
+            );
 
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldNotThrow<Exception>();
@@ -107,20 +109,17 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_execute_user_delegate_async_with_enabled_lambda_return_false()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = false;
+            Context context = new Context { ["ShouldFail"] = false };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true; return Task.FromResult(ResultPrimitive.Good);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(
-                new Exception(),
-                0.4,
-                async (ctx, ct) =>
-                {
-                    return await Task.FromResult((bool)ctx["ShouldFail"]);
-                });
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault(new Exception())
+                    .InjectionRate(0.4)
+                    .EnabledWhen(async (ctx, ct) => await Task.FromResult((bool)ctx["ShouldFail"]))
+            );
 
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldNotThrow<Exception>();
@@ -134,8 +133,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFault_With_Context_Should_not_execute_user_delegate_async_with_default_values()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -145,7 +143,13 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<Context, CancellationToken, Task<Exception>> fault = (ctx, cts) => Task.FromResult(new Exception());
             Func<Context, CancellationToken, Task<double>> injectionRate = (ctx, ct) => Task.FromResult(0.6);
             Func<Context, CancellationToken, Task<bool>> enabled = (ctx, ct) => Task.FromResult(true);
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault<ResultPrimitive>(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldThrowExactly<Exception>();
             executed.Should().BeFalse();
@@ -156,10 +160,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
-            context["Message"] = failureMessage;
-            context["InjectionRate"] = 0.6;
+            Context context = new Context
+            {
+                ["ShouldFail"] = true,
+                ["Message"] = failureMessage,
+                ["InjectionRate"] = 0.6
+            };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -193,7 +199,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context))
                 .ShouldThrowExactly<InvalidOperationException>();
             executed.Should().BeFalse();
@@ -214,7 +225,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
 
             ResultPrimitive fault = ResultPrimitive.Fault;
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, () => true);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.6)
+                    .Enabled()
+            );
+
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -232,7 +248,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
 
             ResultPrimitive fault = ResultPrimitive.Fault;
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.4, () => true);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.4)
+                    .Enabled()
+            );
+
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -242,8 +263,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public async Task InjectFault_With_Context_Enabled_Should_Return_Fault_async()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -256,7 +276,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, enabled);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.6)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -266,8 +291,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public async Task InjectFault_With_Context_Enabled_Should_Not_Return_Fault_async()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = false;
+            Context context = new Context { ["ShouldFail"] = false };
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
                 executed = true;
@@ -280,7 +304,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, 0.6, enabled);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.6)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -290,8 +319,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public async Task InjectFault_With_Context_InjectionRate_Should_Return_Fault_async()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["InjectionRate"] = 0.6;
+            Context context = new Context { ["InjectionRate"] = 0.6 };
 
             Func<Context, Task<ResultPrimitive>> actionAsync = (ctx) =>
             {
@@ -320,7 +348,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult(true);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -360,7 +393,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult(true);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = await policy.ExecuteAsync(actionAsync, context);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -411,7 +449,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return Task.FromResult((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                with.Fault(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             using (CancellationTokenSource cts = new CancellationTokenSource())
             {
                 cts.Cancel();
@@ -468,7 +511,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return Task.FromResult((bool)ctx["ShouldFail"]);
                 };
 
-                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+                var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                    with.Fault(fault)
+                        .InjectionRate(injectionRate)
+                        .EnabledWhen(enabled)
+                );
 
                 policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -522,7 +569,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return Task.FromResult(rate);
                 };
 
-                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+                var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                    with.Fault(fault)
+                        .InjectionRate(injectionRate)
+                        .EnabledWhen(enabled)
+                );
 
                 policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -576,7 +627,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return Task.FromResult(new Exception());
                 };
 
-                var policy = MonkeyPolicy.InjectFaultAsync<ResultPrimitive>(fault, injectionRate, enabled);
+                var policy = MonkeyPolicy.InjectResultAsync<ResultPrimitive>(with =>
+                    with.Fault(fault)
+                        .InjectionRate(injectionRate)
+                        .EnabledWhen(enabled)
+                );
 
                 policy.Awaiting(async x => await x.ExecuteAsync(actionAsync, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
