+++ src/Polly.Contrib.Simmy.Specs/Outcomes/InjectFaultTResultWithOptionsSpecs.cs
@@ -5,13 +5,12 @@ using Polly.Contrib.Simmy.Specs.Helpers;
 using Polly.Contrib.Simmy.Utilities;
 using Xunit;
 
-namespace Polly.Contrib.Simmy.Specs.Fault
+namespace Polly.Contrib.Simmy.Specs.Outcomes
 {
     [Collection(Constants.AmbientContextDependentTestCollection)]
-    [Obsolete]
-    public class InjectFaultTResultSpecs : IDisposable
+    public class InjectFaultTResultWithOptionsSpecs : IDisposable
     {
-        public InjectFaultTResultSpecs()
+        public InjectFaultTResultWithOptionsSpecs()
         {
             ThreadSafeRandom_LockOncePerThread.NextDouble = () => 0.5;
         }
@@ -21,7 +20,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             ThreadSafeRandom_LockOncePerThread.Reset();
         }
 
-        #region Basic Overload, Exception, Context Free
+        #region Basic Overload, Result as Fault, Context Free
         [Fact]
         public void InjectFaultContext_Free_Enabled_Should_not_execute_user_delegate()
         {
@@ -30,7 +29,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Boolean executed = false;
             Func<ResultPrimitive> action = () => { executed = true; return ResultPrimitive.Good; };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, 0.6, () => true);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault<ResultPrimitive>(fault)
+                    .InjectionRate(0.6)
+                    .Enabled()
+            );
+
             policy.Invoking(x => x.Execute(action))
                 .ShouldThrowExactly<Exception>().WithMessage(exceptionMessage);
 
@@ -45,7 +49,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Boolean executed = false;
             Func<ResultPrimitive> action = () => { executed = true; return ResultPrimitive.Good; };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, 0.3, () => true);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault<ResultPrimitive>(fault)
+                    .InjectionRate(0.3)
+                    .Enabled()
+            );
+
             policy.Invoking(x => x.Execute(action))
                 .ShouldNotThrow<Exception>();
 
@@ -53,22 +62,19 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         }
         #endregion
 
-        #region Basic Overload, Exception, With Context
+        #region Basic Overload, Result as Fault, With Context
         [Fact]
         public void InjectFaultWith_Context_Should_not_execute_user_delegate()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(
-                new Exception(),
-                0.6,
-                (ctx, ct) =>
-                {
-                    return ((bool)ctx["ShouldFail"]);
-                });
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault(new Exception())
+                    .InjectionRate(0.6)
+                    .EnabledWhen((ctx, ct) => ((bool)ctx["ShouldFail"]))
+            );
 
             policy.Invoking(x => x.Execute(action, context))
                 .ShouldThrowExactly<Exception>();
@@ -80,17 +86,14 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFaultWith_Context_Should_execute_user_delegate()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(
-                new Exception(),
-                0.4,
-                (ctx, ct) =>
-                {
-                    return ((bool)ctx["ShouldFail"]);
-                });
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault(new Exception())
+                    .InjectionRate(0.4)
+                    .EnabledWhen((ctx, ct) => ((bool)ctx["ShouldFail"]))
+            );
 
             policy.Invoking(x => x.Execute(action, context))
                 .ShouldNotThrow<Exception>();
@@ -102,17 +105,14 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFaultWith_Context_Should_execute_user_delegate_with_enabled_lambda_disabled()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = false;
+            Context context = new Context { ["ShouldFail"] = false };
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(
-                new Exception(),
-                0.6,
-                (ctx, ct) =>
-                {
-                    return ((bool)ctx["ShouldFail"]);
-                });
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault(new Exception())
+                    .InjectionRate(0.6)
+                    .EnabledWhen((ctx, ct) => ((bool)ctx["ShouldFail"]))
+            );
 
             policy.Invoking(x => x.Execute(action, context))
                 .ShouldNotThrow<Exception>();
@@ -132,7 +132,13 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<Context, CancellationToken, Exception> fault = (ctx, ct) => new Exception();
             Func<Context, CancellationToken, double> injectionRate = (ctx, ct) => 0.6;
             Func<Context, CancellationToken, bool> enabled = (ctx, ct) => true;
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, injectionRate, enabled);
+
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault<ResultPrimitive>(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             policy.Invoking(x => x.Execute(action, context))
                 .ShouldThrow<Exception>();
 
@@ -144,10 +150,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
-            context["Message"] = failureMessage;
-            context["InjectionRate"] = 0.6;
+            Context context = new Context
+            {
+                ["ShouldFail"] = true,
+                ["Message"] = failureMessage,
+                ["InjectionRate"] = 0.6
+            };
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
 
             Func<Context, CancellationToken, Exception> fault = (ctx, ct) =>
@@ -175,7 +183,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return ((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Fault(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             policy.Invoking(x => x.Execute(action, context))
                 .ShouldThrowExactly<InvalidOperationException>();
 
@@ -191,7 +204,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<ResultPrimitive> action = () => { executed = true; return ResultPrimitive.Good; };
             ResultPrimitive fault = ResultPrimitive.Fault;
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, 0.6, () => true);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.6)
+                    .Enabled()
+            );
+
             ResultPrimitive response = policy.Execute(action);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -204,7 +222,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
             Func<ResultPrimitive> action = () => { executed = true; return ResultPrimitive.Good; };
             ResultPrimitive fault = ResultPrimitive.Fault;
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, 0.4, () => true);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.4)
+                    .Enabled()
+            );
+
             ResultPrimitive response = policy.Execute(action);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -214,8 +237,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFaultWith_Context_Enabled_Should_Return_Fault()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
+            Context context = new Context { ["ShouldFail"] = true };
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
             ResultPrimitive fault = ResultPrimitive.Fault;
             Func<Context, CancellationToken, bool> enabled = (ctx, ct) =>
@@ -223,7 +245,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return ((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, 0.6, enabled);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.6)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = policy.Execute(action, context);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -233,8 +260,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFaultWith_Context_Enabled_Should_Not_Return_Fault()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = false;
+            Context context = new Context { ["ShouldFail"] = false };
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
             ResultPrimitive fault = ResultPrimitive.Fault;
             Func<Context, CancellationToken, bool> enabled = (ctx, ct) =>
@@ -242,7 +268,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return ((bool)ctx["ShouldFail"]);
             };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, 0.6, enabled);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(0.4)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = policy.Execute(action, context);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -252,8 +283,7 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         public void InjectFaultWith_Context_InjectionRate_Should_Return_Fault()
         {
             Boolean executed = false;
-            Context context = new Context();
-            context["InjectionRate"] = 0.6;
+            Context context = new Context { ["InjectionRate"] = 0.6 };
 
             Func<Context, ResultPrimitive> action = (ctx) => { executed = true; return ResultPrimitive.Good; };
             Func<Context, CancellationToken, ResultPrimitive> fault = (ctx, ct) =>
@@ -276,7 +306,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return true;
             };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = policy.Execute(action, context);
             response.Should().Be(ResultPrimitive.Fault);
             executed.Should().BeFalse();
@@ -310,7 +345,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                 return true;
             };
 
-            var policy = MonkeyPolicy.InjectFault<ResultPrimitive>(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectResult<ResultPrimitive>(with =>
+                with.Result(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             ResultPrimitive response = policy.Execute(action, context);
             response.Should().Be(ResultPrimitive.Good);
             executed.Should().BeTrue();
@@ -324,10 +364,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
-            context["Message"] = failureMessage;
-            context["InjectionRate"] = 0.6;
+            Context context = new Context
+            {
+                ["ShouldFail"] = true, ["Message"] = failureMessage, ["InjectionRate"] = 0.6
+            };
 
             Func<Context, CancellationToken, Exception> fault = (ctx, cts) =>
             {
@@ -358,7 +398,12 @@ namespace Polly.Contrib.Simmy.Specs.Fault
 
             Func<Context, CancellationToken, ResultPrimitive> action = (ctx, ct) => { executed = true; return ResultPrimitive.Good; };
 
-            var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
+            var policy = MonkeyPolicy.InjectException(with =>
+                with.Fault(fault)
+                    .InjectionRate(injectionRate)
+                    .EnabledWhen(enabled)
+            );
+
             using (CancellationTokenSource cts = new CancellationTokenSource())
             {
                 cts.Cancel();
@@ -375,10 +420,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
-            context["Message"] = failureMessage;
-            context["InjectionRate"] = 0.6;
+            Context context = new Context
+            {
+                ["ShouldFail"] = true, ["Message"] = failureMessage, ["InjectionRate"] = 0.6
+            };
 
             Func<Context, CancellationToken, Exception> fault = (ctx, cts) =>
             {
@@ -412,7 +457,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return (bool)ctx["ShouldFail"];
                 };
 
-                var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
+                var policy = MonkeyPolicy.InjectException(with =>
+                    with.Fault(fault)
+                        .InjectionRate(injectionRate)
+                        .EnabledWhen(enabled)
+                );
 
                 policy.Invoking(x => x.Execute(action, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -426,10 +475,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
-            context["Message"] = failureMessage;
-            context["InjectionRate"] = 0.6;
+            Context context = new Context
+            {
+                ["ShouldFail"] = true, ["Message"] = failureMessage, ["InjectionRate"] = 0.6
+            };
 
             Func<Context, CancellationToken, Exception> fault = (ctx, cts) =>
             {
@@ -463,7 +512,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
                     return rate;
                 };
 
-                var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
+                var policy = MonkeyPolicy.InjectException(with =>
+                    with.Fault(fault)
+                        .InjectionRate(injectionRate)
+                        .EnabledWhen(enabled)
+                );
 
                 policy.Invoking(x => x.Execute(action, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
@@ -477,10 +530,10 @@ namespace Polly.Contrib.Simmy.Specs.Fault
         {
             string failureMessage = "Failure Message";
             Boolean executed = false;
-            Context context = new Context();
-            context["ShouldFail"] = true;
-            context["Message"] = failureMessage;
-            context["InjectionRate"] = 0.6;
+            Context context = new Context
+            {
+                ["ShouldFail"] = true, ["Message"] = failureMessage, ["InjectionRate"] = 0.6
+            };
 
             using (CancellationTokenSource cts = new CancellationTokenSource())
             {
@@ -514,7 +567,11 @@ namespace Polly.Contrib.Simmy.Specs.Fault
 
                 Func<Context, CancellationToken, ResultPrimitive> action = (ctx, ct) => { executed = true; return ResultPrimitive.Good; };
 
-                var policy = MonkeyPolicy.InjectFault(fault, injectionRate, enabled);
+                var policy = MonkeyPolicy.InjectException(with =>
+                    with.Fault(fault)
+                        .InjectionRate(injectionRate)
+                        .EnabledWhen(enabled)
+                );
 
                 policy.Invoking(x => x.Execute(action, context, cts.Token))
                     .ShouldThrow<OperationCanceledException>();
