+++ src/Polly.Contrib.Simmy/Behavior/AsyncInjectBehaviourPolicy.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Polly.Contrib.Simmy.Behavior.Options;
 
 namespace Polly.Contrib.Simmy.Behavior
 {
