+++ src/Polly.Contrib.Simmy/Behavior/InjectBehaviourAsyncOptions.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Polly.Contrib.Simmy.Behavior;
 
-namespace Polly.Contrib.Simmy.Behavior.Options
+namespace Polly.Contrib.Simmy
 {
     /// <summary>
     /// Options used to configure an <see cref="AsyncInjectBehaviourPolicy"/>
