+++ src/Polly.Contrib.Simmy/Behavior/InjectBehaviourAsyncOptionsExtensions.cs
@@ -1,19 +1,20 @@
 ﻿using System;
 using System.Threading;
+using System.Threading.Tasks;
 
-namespace Polly.Contrib.Simmy.Behavior.Options
+namespace Polly.Contrib.Simmy
 {
     /// <summary>
-    /// Allows configuration of behaviour for synchronous monkey behaviour-injection policies.
+    /// Allows configuration of behaviour for asynchronous monkey behaviour-injection policies.
     /// </summary>
-    public static class InjectBehaviourOptionsExtensions
+    public static class InjectBehaviourAsyncOptionsExtensions
     {
         /// <summary>
         /// Configure behaviour to inject with the monkey policy.
         /// </summary>
         /// <param name="options">The configuration object.</param>
         /// <param name="behaviour">A delegate representing the behaviour to inject.</param>
-        public static InjectBehaviourOptions Behaviour(this InjectBehaviourOptions options, Action behaviour) =>
+        public static InjectBehaviourAsyncOptions Behaviour(this InjectBehaviourAsyncOptions options, Func<Task> behaviour) =>
             Behaviour(options, (_, __) => behaviour());
 
         /// <summary>
@@ -21,7 +22,7 @@ namespace Polly.Contrib.Simmy.Behavior.Options
         /// </summary>
         /// <param name="options">The configuration object.</param>
         /// <param name="behaviour">A delegate representing the behaviour to inject.</param>
-        public static InjectBehaviourOptions Behaviour(this InjectBehaviourOptions options, Action<Context, CancellationToken> behaviour)
+        public static InjectBehaviourAsyncOptions Behaviour(this InjectBehaviourAsyncOptions options, Func<Context, CancellationToken, Task> behaviour)
         {
             options.Behaviour = behaviour;
             return options;
