+++ src/Polly.Contrib.Simmy/Behavior/InjectBehaviourOptionsExtensions.cs
@@ -1,20 +1,19 @@
 ﻿using System;
 using System.Threading;
-using System.Threading.Tasks;
 
-namespace Polly.Contrib.Simmy.Behavior.Options
+namespace Polly.Contrib.Simmy
 {
     /// <summary>
-    /// Allows configuration of behaviour for asynchronous monkey behaviour-injection policies.
+    /// Allows configuration of behaviour for synchronous monkey behaviour-injection policies.
     /// </summary>
-    public static class InjectBehaviourAsyncOptionsExtensions
+    public static class InjectBehaviourOptionsExtensions
     {
         /// <summary>
         /// Configure behaviour to inject with the monkey policy.
         /// </summary>
         /// <param name="options">The configuration object.</param>
         /// <param name="behaviour">A delegate representing the behaviour to inject.</param>
-        public static InjectBehaviourAsyncOptions Behaviour(this InjectBehaviourAsyncOptions options, Func<Task> behaviour) =>
+        public static InjectBehaviourOptions Behaviour(this InjectBehaviourOptions options, Action behaviour) =>
             Behaviour(options, (_, __) => behaviour());
 
         /// <summary>
@@ -22,7 +21,7 @@ namespace Polly.Contrib.Simmy.Behavior.Options
         /// </summary>
         /// <param name="options">The configuration object.</param>
         /// <param name="behaviour">A delegate representing the behaviour to inject.</param>
-        public static InjectBehaviourAsyncOptions Behaviour(this InjectBehaviourAsyncOptions options, Func<Context, CancellationToken, Task> behaviour)
+        public static InjectBehaviourOptions Behaviour(this InjectBehaviourOptions options, Action<Context, CancellationToken> behaviour)
         {
             options.Behaviour = behaviour;
             return options;
