+++ src/Polly.Contrib.Simmy/Behavior/InjectBehaviourTResultOptionsSyntax.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Polly.Contrib.Simmy.Behavior;
-using Polly.Contrib.Simmy.Behavior.Options;
 
 namespace Polly.Contrib.Simmy
 {
@@ -18,7 +17,7 @@ namespace Polly.Contrib.Simmy
         public static InjectBehaviourPolicy<TResult> InjectBehaviour<TResult>(Action<InjectBehaviourOptions> configureOptions)
         {
             var options = new InjectBehaviourOptions();
-            configureOptions.Invoke(options);
+            configureOptions(options);
 
             if (options.Behaviour == null) throw new ArgumentNullException(nameof(options.Behaviour));
             if (options.InjectionRate == null) throw new ArgumentNullException(nameof(options.InjectionRate));
