+++ src/Polly.Contrib.Simmy/Latency/AsyncInjectLatencyPolicy.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Polly.Contrib.Simmy.Latency.Options;
 using Polly.Utilities;
 
 namespace Polly.Contrib.Simmy.Latency
