+++ src/Polly.Contrib.Simmy/Outcomes/AsyncInjectFaultOptionsSyntax.cs
@@ -1,6 +1,5 @@
 ﻿using System;
-using Polly.Contrib.Simmy.Fault;
-using Polly.Contrib.Simmy.Fault.Options;
+using Polly.Contrib.Simmy.Outcomes;
 
 namespace Polly.Contrib.Simmy
 {
@@ -15,10 +14,10 @@ namespace Polly.Contrib.Simmy
         /// </summary>
         /// <param name="configureOptions">A callback to configure policy options.</param>
         /// <returns>The policy instance.</returns>
-        public static AsyncInjectOutcomePolicy InjectFaultAsync(Action<InjectFaultAsyncOptions<Exception>> configureOptions)
+        public static AsyncInjectOutcomePolicy InjectExceptionAsync(Action<InjectFaultAsyncOptions<Exception>> configureOptions)
         {
             var options = new InjectFaultAsyncOptions<Exception>();
-            configureOptions.Invoke(options);
+            configureOptions(options);
 
             if (options.Outcome == null) throw new ArgumentNullException(nameof(options.Outcome));
             if (options.InjectionRate == null) throw new ArgumentNullException(nameof(options.InjectionRate));
@@ -28,15 +27,15 @@ namespace Polly.Contrib.Simmy
         }
 
         /// <summary>
-        /// Builds an <see cref="AsyncInjectOutcomePolicy"/> which injects a fault if <paramref name="configureOptions.Enabled"/> returns true and
+        /// Builds an <see cref="AsyncInjectOutcomePolicy"/> which injects a result if <paramref name="configureOptions.Enabled"/> returns true and
         /// a random number is within range of <paramref name="configureOptions.InjectionRate"/>.
         /// </summary>
         /// <param name="configureOptions">A callback to configure policy options.</param>
         /// <returns>The policy instance.</returns>
-        public static AsyncInjectOutcomePolicy<TResult> InjectFaultAsync<TResult>(Action<InjectFaultAsyncOptions<TResult>> configureOptions)
+        public static AsyncInjectOutcomePolicy<TResult> InjectResultAsync<TResult>(Action<InjectFaultAsyncOptions<TResult>> configureOptions)
         {
             var options = new InjectFaultAsyncOptions<TResult>();
-            configureOptions.Invoke(options);
+            configureOptions(options);
 
             if (options.Outcome == null) throw new ArgumentNullException(nameof(options.Outcome));
             if (options.InjectionRate == null) throw new ArgumentNullException(nameof(options.InjectionRate));
@@ -46,15 +45,15 @@ namespace Polly.Contrib.Simmy
         }
 
         /// <summary>
-        /// Builds an <see cref="AsyncInjectOutcomePolicy"/> which injects a fault if <paramref name="configureOptions.Enabled"/> returns true and
+        /// Builds an <see cref="AsyncInjectOutcomePolicy"/> which injects a fault as result if  <paramref name="configureOptions.Enabled"/> returns true and
         /// a random number is within range of <paramref name="configureOptions.InjectionRate"/>.
         /// </summary>
         /// <param name="configureOptions">A callback to configure policy options.</param>
         /// <returns>The policy instance.</returns>
-        public static AsyncInjectOutcomePolicy<TResult> InjectFaultAsync<TResult>(Action<InjectFaultAsyncOptions<Exception>> configureOptions)
+        public static AsyncInjectOutcomePolicy<TResult> InjectResultAsync<TResult>(Action<InjectFaultAsyncOptions<Exception>> configureOptions)
         {
             var options = new InjectFaultAsyncOptions<Exception>();
-            configureOptions.Invoke(options);
+            configureOptions(options);
 
             if (options.Outcome == null) throw new ArgumentNullException(nameof(options.Outcome));
             if (options.InjectionRate == null) throw new ArgumentNullException(nameof(options.InjectionRate));
