+++ src/Polly.Contrib.Simmy/Outcomes/InjectFaultOptions.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Threading;
+using Polly.Contrib.Simmy.Outcomes;
 
-namespace Polly.Contrib.Simmy.Fault.Options
+namespace Polly.Contrib.Simmy
 {
     /// <summary>
     /// Options used to configure an <see cref="InjectOutcomePolicy"/>
