+++ src/Polly.Contrib.Simmy/Outcomes/InjectFaultOptionsExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading;
 
-namespace Polly.Contrib.Simmy.Fault.Options
+namespace Polly.Contrib.Simmy
 {
     /// <summary>
     /// Allows configuration of fault for hronous monkey fault-injection policies.
@@ -47,7 +47,7 @@ namespace Polly.Contrib.Simmy.Fault.Options
         }
 
         /// <summary>
-        /// Configure fault to inject with the monkey policy.
+        /// Configure result to inject with the monkey policy.
         /// </summary>
         /// <param name="options">The configuration object.</param>
         /// <param name="result">The result to inject</param>
@@ -55,7 +55,7 @@ namespace Polly.Contrib.Simmy.Fault.Options
             Result(options, (_, __) => result);
 
         /// <summary>
-        /// Configure fault to inject with the monkey policy.
+        /// Configure result to inject with the monkey policy.
         /// </summary>
         /// <param name="options">The configuration object.</param>
         /// <param name="result">A delegate representing the result to inject.</param>
